ml-service/requirements.txt
flask
tensorflow-cpu==2.16.1
pandas
scikit-learn
numpy
scipy
werkzeug
